Rename deref.rs to containers.rs
